Add Furigana lookup functionality using MeCab, allowing the use of Gemma 2 2b, which works very fast offline on many devices, while maintaining furigana accurracy.
