+++ config/current_copy_mode.json
@@ -1,1 +1,1 @@
-{"mode": 1}
+{"mode": 0}
